requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.3
idna==3.11
kombu==5.6.2
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
prompt_toolkit==3.0.52
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
